requirements.txt

Flask==3.0.3
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
Flask-Marshmallow==1.2.1
marshmallow==3.21.3
marshmallow-sqlalchemy==0.29.0
PyJWT==2.9.0
python-dotenv==1.0.1
pytest==8.3.3
werkzeug==3.0.4
